+++ include/Utils/Rotation.h
@@ -117,7 +117,7 @@ public:
     /// \return The axis
     ///
     Vector3d axe(
-        unsigned int idx) const;
+        size_t idx) const;
 
     ///
     /// \brief set the Rotation from a spatial transform
