+++ src/InternalForces/Actuators/ActuatorLinear.cpp
@@ -27,7 +27,7 @@ internal_forces::actuator::ActuatorLinear::ActuatorLinear(
     int direction,
     const utils::Scalar& T0,
     const utils::Scalar& slope,
-    unsigned int dofIdx) :
+    size_t dofIdx) :
     Actuator(direction, dofIdx),
     m_m(std::make_shared<utils::Scalar>(slope)),
     m_b(std::make_shared<utils::Scalar>(T0))
@@ -39,7 +39,7 @@ internal_forces::actuator::ActuatorLinear::ActuatorLinear(
     int direction,
     const utils::Scalar& T0,
     const utils::Scalar& slope,
-    unsigned int dofIdx,
+    size_t dofIdx,
     const utils::String &jointName) :
     Actuator(direction, dofIdx, jointName),
     m_m(std::make_shared<utils::Scalar>(slope)),
@@ -78,7 +78,7 @@ utils::Scalar internal_forces::actuator::ActuatorLinear::torqueMax()
 utils::Scalar internal_forces::actuator::ActuatorLinear::torqueMax(
     const rigidbody::GeneralizedCoordinates &Q) const
 {
-    return (Q[*m_dofIdx]*180/M_PI) * *m_m + *m_b;
+    return (Q[static_cast<unsigned int>(*m_dofIdx)]*180/M_PI) * *m_m + *m_b;
 }
 
 void internal_forces::actuator::ActuatorLinear::setType()
